support writeback requests from L1 caches
SinkC now handles ProbeAcks with M_FLUSH_WB param. The MSHR
is queued with control1 high. Based on the state of the cache line
in L2, the MSHR may or may not probe caches of other clients. Probes
happen in such a way as to not invalidate the cache lines of other clients.
Probes simply downgrade.
The state of the L2 in case of a writeback is either TRUNK or TIP. TRUNK
is when a dirty L1 is the requester of a writeback. TIP is in every other situation.